+++ app/gui/src/components/ProtectedLayout.vue
@@ -16,9 +16,11 @@ import LocalStorage from '#/utilities/LocalStorage'
 import { DASHBOARD_PATH, LOGIN_PATH, RESTORE_USER_PATH } from '$/appUtils'
 import { useUserAgreements } from '$/composables/userAgreements'
 import { useAuth, type AuthStore } from '$/providers/auth'
+import { useFeatureFlag } from '$/providers/featureFlags'
 import { useSession } from '$/providers/session'
 import { useText } from '$/providers/text'
 import type { DataLoader } from '$/router'
+import { useAppClass } from '@/providers/appClass'
 import { Dialog, reactComponent, ResultComponent } from '@/util/react'
 import * as vueQuery from '@tanstack/vue-query'
 import { useQueryClient } from '@tanstack/vue-query'
@@ -129,6 +131,10 @@ const queryClient = useQueryClient()
 const text = useText()
 const EnsoDevtools = reactComponent(EnsoDevToolsReact)
 const ReactQueryDevtools = reactComponent(ReactQueryDevtoolsReact)
+
+// Needed by devtools - act on feature flag changes
+const debugHoverAreas = useFeatureFlag('debugHoverAreas')
+useAppClass(() => ({ debugHoverAreas: debugHoverAreas.value }))
 
 const allowed = computed(() => routeAllowed(route, auth))
 watch(