+++ CHANGELOG.md
@@ -14,6 +14,7 @@
   is included][14028]
 - [Function docs in autocomplete in table expressions][14059]
 - [Many CLI arguments removed][14069]
+- [Multiple opened projects' tabs are now allowed][14215]
 
 [13685]: https://github.com/enso-org/enso/pull/13685
 [13658]: https://github.com/enso-org/enso/pull/13658
@@ -26,6 +27,7 @@
 [13976]: https://github.com/enso-org/enso/pull/13976
 [14059]: https://github.com/enso-org/enso/pull/14059
 [14069]: https://github.com/enso-org/enso/pull/14069
+[14215]: https://github.com/enso-org/enso/pull/14215
 
 #### Enso Standard Library
 
@@ -42,6 +44,8 @@
 - [Add Text_Column.proper and Rename Case.Title->Case.Proper.][14184]
 - [Snowflake stage support for reading and writing files.][14210]
 - [Snowflake file format and copy into support.][14221]
+- [Snowflake bulk load API.][14230]
+- [Add error_on_missing_columns to more methods.][14236]
 
 [13769]: https://github.com/enso-org/enso/pull/13769
 [14026]: https://github.com/enso-org/enso/pull/14026
@@ -55,6 +59,8 @@
 [14184]: https://github.com/enso-org/enso/pull/14184
 [14210]: https://github.com/enso-org/enso/pull/14210
 [14221]: https://github.com/enso-org/enso/pull/14221
+[14230]: https://github.com/enso-org/enso/pull/14230
+[14236]: https://github.com/enso-org/enso/pull/14236
 
 #### Enso Language & Runtime
 
@@ -70,7 +76,7 @@
 - [Autocompletion for table expression builtin functions in table
   expressions][13914]
 - [Autocompletion for table expression operators and operands][13917]
-- [Update to GraalVM 25][14019]
+- [Update to GraalVM 25.0.1][14233]
 
 [14133]: https://github.com/enso-org/enso/pull/14133
 [14017]: https://github.com/enso-org/enso/pull/14017
@@ -83,7 +89,7 @@
 [13178]: https://github.com/enso-org/enso/pull/13178
 [13914]: https://github.com/enso-org/enso/pull/13914
 [13917]: https://github.com/enso-org/enso/pull/13917
-[14019]: https://github.com/enso-org/enso/pull/14019
+[14233]: https://github.com/enso-org/enso/pull/14233
 
 # Enso 2025.2